+++ builders/obfuscate/options.d.ts
@@ -1,13 +1,13 @@
 import { IOptions } from 'javascript-obfuscator/typings/src/interfaces/options/IOptions';
 
-interface ObfuscateCommandOptions extends IOptions {
+export interface ObfuscateCommandOptions extends IOptions {
   /**
    * The files which should get obfuscated.
    */
   files: Array<JsFiles>;
 }
 
-interface JsFiles {
+export interface JsFiles {
   fileExtensions: string[];
   filePatternsToExclude: string[];
   input: string;